debug

+++ task.py
@@ -3921,7 +3921,7 @@ def train_v13(config,
 
   def _accumulate_adv_gradients(source, target):
     adv_domain = tf.math.divide(source["domain"][0] + \
-      tf.reshape(tf.random.categorical(tf.expand_dims(tf.math.log([1.0/(domain_num-1)]*domain_num),0),1),[]) + 1, domain_num)
+      tf.cast(tf.reshape(tf.random.categorical(tf.expand_dims(tf.math.log([1.0/(domain_num-1)]*domain_num),0),1),[]) + 1, domain_num),tf.int32)
     source["domain"] = tf.tile(tf.expand_dims(adv_domain,0), source["domain"].shape)
     target["domain"] = tf.tile(tf.expand_dims(adv_domain,0), target["domain"].shape)
     outputs, _ = model(
